llcppsigfetch:split test directory
