+++ chore/_xtool/llcppsigfetch/parse/cvt_test/decl_test/class_test/llgo.expect
@@ -1,0 +1,169 @@
+#stdout
+TestClassDecl Case 1:
+{
+	"temp.h":	{
+		"decls":	[{
+				"Loc":	{
+					"File":	"temp.h"
+				},
+				"Doc":	{
+					"List":	[]
+				},
+				"Parent":	{
+					"Name":	"A"
+				},
+				"Name":	{
+					"Name":	"A"
+				},
+				"Type":	{
+					"Tag":	3,
+					"Fields":	{
+						"List":	[{
+								"Type":	{
+									"Kind":	6,
+									"Flags":	0
+								},
+								"Doc":	{
+									"List":	[]
+								},
+								"Comment":	{
+									"List":	[]
+								},
+								"Names":	[{
+										"Name":	"a"
+									}]
+							}, {
+								"Type":	{
+									"Kind":	6,
+									"Flags":	0
+								},
+								"Doc":	{
+									"List":	[]
+								},
+								"Comment":	{
+									"List":	[]
+								},
+								"Names":	[{
+										"Name":	"b"
+									}]
+							}]
+					},
+					"Methods":	[]
+				}
+			}],
+		"includes":	[],
+		"macros":	[]
+	}
+}
+
+TestClassDecl Case 2:
+{
+	"temp.h":	{
+		"decls":	[{
+				"Loc":	{
+					"File":	"temp.h"
+				},
+				"Doc":	{
+					"List":	[]
+				},
+				"Parent":	{
+					"Name":	"A"
+				},
+				"Name":	{
+					"Name":	"A"
+				},
+				"Type":	{
+					"Tag":	3,
+					"Fields":	{
+						"List":	[{
+								"Type":	{
+									"Kind":	6,
+									"Flags":	0
+								},
+								"Doc":	{
+									"List":	[]
+								},
+								"Comment":	{
+									"List":	[]
+								},
+								"Names":	[{
+										"Name":	"a"
+									}]
+							}, {
+								"Type":	{
+									"Kind":	6,
+									"Flags":	0
+								},
+								"Doc":	{
+									"List":	[]
+								},
+								"Comment":	{
+									"List":	[]
+								},
+								"Names":	[{
+										"Name":	"b"
+									}]
+							}]
+					},
+					"Methods":	[{
+							"Loc":	{
+								"File":	"temp.h"
+							},
+							"Doc":	{
+								"List":	[]
+							},
+							"Parent":	{
+								"Name":	"A"
+							},
+							"Name":	{
+								"Name":	"foo"
+							},
+							"Type":	{
+								"Params":	{
+									"List":	[{
+											"Type":	{
+												"Kind":	6,
+												"Flags":	0
+											},
+											"Doc":	{
+												"List":	[]
+											},
+											"Comment":	{
+												"List":	[]
+											},
+											"Names":	[{
+													"Name":	"a"
+												}]
+										}, {
+											"Type":	{
+												"Kind":	8,
+												"Flags":	16
+											},
+											"Doc":	{
+												"List":	[]
+											},
+											"Comment":	{
+												"List":	[]
+											},
+											"Names":	[{
+													"Name":	"b"
+												}]
+										}]
+								},
+								"Ret":	{
+									"Kind":	8,
+									"Flags":	0
+								}
+							}
+						}]
+				}
+			}],
+		"includes":	[],
+		"macros":	[]
+	}
+}
+
+
+#stderr
+
+#exit 0
