+++ chore/_xtool/llcppsigfetch/parse/cvt_test/decl_test/enum_test/llgo.expect
@@ -1,0 +1,140 @@
+#stdout
+TestEnumDecl Case 1:
+{
+	"temp.h":	{
+		"decls":	[{
+				"Loc":	{
+					"File":	"temp.h"
+				},
+				"Doc":	{
+					"List":	[]
+				},
+				"Parent":	null,
+				"Name":	{
+					"Name":	"Foo"
+				},
+				"Items":	[{
+						"Name":	{
+							"Name":	"a"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"0"
+						}
+					}, {
+						"Name":	{
+							"Name":	"b"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"1"
+						}
+					}, {
+						"Name":	{
+							"Name":	"c"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"2"
+						}
+					}]
+			}],
+		"includes":	[],
+		"macros":	[]
+	}
+}
+
+TestEnumDecl Case 2:
+{
+	"temp.h":	{
+		"decls":	[{
+				"Loc":	{
+					"File":	"temp.h"
+				},
+				"Doc":	{
+					"List":	[]
+				},
+				"Parent":	null,
+				"Name":	{
+					"Name":	"Foo"
+				},
+				"Items":	[{
+						"Name":	{
+							"Name":	"a"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"1"
+						}
+					}, {
+						"Name":	{
+							"Name":	"b"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"2"
+						}
+					}, {
+						"Name":	{
+							"Name":	"c"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"4"
+						}
+					}]
+			}],
+		"includes":	[],
+		"macros":	[]
+	}
+}
+
+TestEnumDecl Case 3:
+{
+	"temp.h":	{
+		"decls":	[{
+				"Loc":	{
+					"File":	"temp.h"
+				},
+				"Doc":	{
+					"List":	[]
+				},
+				"Parent":	null,
+				"Name":	{
+					"Name":	"Foo"
+				},
+				"Items":	[{
+						"Name":	{
+							"Name":	"a"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"1"
+						}
+					}, {
+						"Name":	{
+							"Name":	"b"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"2"
+						}
+					}, {
+						"Name":	{
+							"Name":	"c"
+						},
+						"Value":	{
+							"Kind":	0,
+							"Value":	"3"
+						}
+					}]
+			}],
+		"includes":	[],
+		"macros":	[]
+	}
+}
+
+
+#stderr
+
+#exit 0
