+++ chore/_xtool/llcppsigfetch/parse/cvt_test/decl_test/scope_test/scope.go
@@ -1,0 +1,30 @@
+package main
+
+import test "github.com/goplus/llgo/chore/_xtool/llcppsigfetch/parse/cvt_test"
+
+func main() {
+	TestScope()
+}
+
+func TestScope() {
+	testCases := []string{
+		`void foo();`,
+		`namespace a {
+			 void foo(); 
+		 }`,
+		`namespace a {
+			namespace b {
+				void foo(); 
+			}
+	   	 }`,
+		`class a {
+			void foo(); 
+		 };`,
+		`namespace a {
+		 class b {
+			void foo(); 
+		 };
+	   	 }`,
+	}
+	test.RunTest("TestScope", testCases)
+}
