+++ chore/_xtool/llcppsigfetch/parse/cvt_test/preprocess_test/llgo.expect
@@ -2,7 +2,6 @@
 TestDefine Case 1:
 {
 	"temp.h":	{
-		"path":	"temp.h",
 		"decls":	[],
 		"includes":	[],
 		"macros":	[{
@@ -18,7 +17,6 @@ TestDefine Case 1:
 TestDefine Case 2:
 {
 	"temp.h":	{
-		"path":	"temp.h",
 		"decls":	[],
 		"includes":	[],
 		"macros":	[{
@@ -37,7 +35,6 @@ TestDefine Case 2:
 TestDefine Case 3:
 {
 	"temp.h":	{
-		"path":	"temp.h",
 		"decls":	[],
 		"includes":	[],
 		"macros":	[{
@@ -89,7 +86,6 @@ TestDefine Case 3:
 TestInclude Case 1:
 {
 	"temp.h":	{
-		"path":	"temp.h",
 		"decls":	[],
 		"includes":	[{
 				"Path":	"foo.h"
